+++ soc/ti/k3/CMakeLists.txt
@@ -1,4 +1,4 @@
 # Copyright (c) 2024 Nordic Semiconductor ASA
 # SPDX-License-Identifier: Apache-2.0
 
-add_subdirectory(${SOC_SERIES})
+add_subdirectory(./common)
