+++ soc/ti/k3/Kconfig
@@ -1,8 +1,19 @@
+# Copyright (c) 2023 Enphase Energy
 # Copyright (c) 2024 Nordic Semiconductor ASA
+# Copyright (c) 2026 Siemens Mobility GmbH
 # SPDX-License-Identifier: Apache-2.0
 
 if SOC_FAMILY_TI_K3
 
 rsource "*/Kconfig"
+
+config SOC_PART_NUMBER
+	default "AM2434" if SOC_AM2434_M4 || SOC_AM2434_R5F0_0
+	default "AM6232" if SOC_AM6232_A53 || SOC_AM6232_M4
+	default "AM6234" if SOC_AM6234_A53 || SOC_AM6234_M4
+	default "AM6254" if SOC_AM6254_A53 || SOC_AM6254_M4
+	default "AM62L3" if SOC_AM62L3_A53
+	default "AM6442" if SOC_AM6442_M4 || SOC_AM6442_R5F0_0 || SOC_AM6442_R5F0_1 || SOC_AM6442_R5F1_0 || SOC_AM6442_R5F1_1
+	default "J721e" if SOC_J721E_MAIN_R5F0_0
 
 endif # SOC_FAMILY_TI_K3
