+++ soc/ti/k3/am62lx/Kconfig.defconfig
@@ -1,0 +1,27 @@
+# Copyright (c) 2023 Enphase Energy
+# Copyright (c) 2026 Siemens Mobility GmbH
+# SPDX-License-Identifier: Apache-2.0
+
+if SOC_SERIES_AM62LX
+
+config NUM_IRQS
+	default 992 if SOC_SERIES_AM62LX_A53
+
+config SYS_CLOCK_HW_CYCLES_PER_SEC
+	default 200000000 if SOC_SERIES_AM62LX_A53
+
+if SERIAL
+
+config UART_NS16550
+	default y
+
+config UART_NS16550_TI_K3
+	default y if SOC_SERIES_AM62LX_A53
+
+choice UART_NS16550_VARIANT
+	default UART_NS16550_VARIANT_NS16750
+endchoice
+
+endif # SERIAL
+
+endif # SOC_SERIES_AM62LX
