+++ soc/ti/k3/am62x/Kconfig
@@ -1,0 +1,23 @@
+# Copyright (c) 2023 Enphase Energy
+# Copyright (c) 2026 Siemens Mobility GmbH
+# SPDX-License-Identifier: Apache-2.0
+
+config SOC_SERIES_AM62X_A53
+	select ARM64
+	select CPU_CORTEX_A53
+	select ARM_ARCH_TIMER if SYS_CLOCK_EXISTS
+	select SOC_EARLY_INIT_HOOK
+
+config SOC_SERIES_AM62X_M4
+	select ARM
+	select CPU_CORTEX_M4
+	select CPU_HAS_FPU
+	select CPU_CORTEX_M_HAS_SYSTICK
+	select DYNAMIC_INTERRUPTS
+	select CPU_CORTEX_M_HAS_DWT
+	select OPENAMP_RSC_TABLE
+	select UART_NS16550_ACCESS_WORD_ONLY
+	select EXTERNAL_ADDRESS_TRANSLATION
+	select MM_DRV
+	select MM_TI_RAT
+	select SOC_EARLY_INIT_HOOK
