+++ soc/ti/k3/am64x/Kconfig
@@ -1,0 +1,29 @@
+# Copyright (c) 2023 Enphase Energy
+# Copyright (c) 2026 Siemens Mobility GmbH
+# SPDX-License-Identifier: Apache-2.0
+
+config SOC_SERIES_AM64X_M4
+	select ARM
+	select CPU_CORTEX_M4
+	select CPU_HAS_FPU
+	select CPU_CORTEX_M_HAS_SYSTICK
+	select DYNAMIC_INTERRUPTS
+	select CPU_CORTEX_M_HAS_DWT
+	select OPENAMP_RSC_TABLE
+	select UART_NS16550_ACCESS_WORD_ONLY
+	select EXTERNAL_ADDRESS_TRANSLATION
+	select MM_DRV
+	select MM_TI_RAT
+	select SOC_EARLY_INIT_HOOK
+
+config SOC_SERIES_AM64X_R5
+	select ARM
+	select CPU_CORTEX_R5
+	select CPU_HAS_ARM_MPU
+	select VFP_SP_D16
+	select ARM_CUSTOM_INTERRUPT_CONTROLLER
+	select TI_VIM
+	select TI_DM_TIMER
+	select OPENAMP_RSC_TABLE
+	select UART_NS16550_ACCESS_WORD_ONLY if UART_NS16550
+	select SOC_EARLY_INIT_HOOK
