--- soc/ti/k3/am6x/Kconfig
@@ -1,49 +1,0 @@
-# Copyright (c) 2023 Enphase Energy
-# SPDX-License-Identifier: Apache-2.0
-
-config SOC_SERIES_AM6X_A53
-	select ARM64
-	select CPU_CORTEX_A53
-	select ARM_ARCH_TIMER if SYS_CLOCK_EXISTS
-	select SOC_EARLY_INIT_HOOK
-
-config SOC_SERIES_AM6X_M4
-	select ARM
-	select CPU_CORTEX_M4
-	select CPU_HAS_FPU
-	select CPU_CORTEX_M_HAS_SYSTICK
-	select DYNAMIC_INTERRUPTS
-	select CPU_CORTEX_M_HAS_DWT
-	select OPENAMP_RSC_TABLE
-	select UART_NS16550_ACCESS_WORD_ONLY
-	select EXTERNAL_ADDRESS_TRANSLATION
-	select MM_DRV
-	select MM_TI_RAT
-	select SOC_EARLY_INIT_HOOK
-
-config SOC_SERIES_AM6X_R5
-	select ARM
-	select CPU_CORTEX_R5
-	select CPU_HAS_ARM_MPU
-	select VFP_SP_D16
-	select ARM_CUSTOM_INTERRUPT_CONTROLLER
-	select TI_VIM
-	select TI_DM_TIMER
-	select OPENAMP_RSC_TABLE
-	select UART_NS16550_ACCESS_WORD_ONLY if UART_NS16550
-	select SOC_EARLY_INIT_HOOK
-
-config SOC_PART_NUMBER
-	default "AM2434" if SOC_AM2434_M4 || SOC_AM2434_R5F0_0
-	default "AM6232" if SOC_AM6232_A53
-	default "AM6232" if SOC_AM6232_M4
-	default "AM6234" if SOC_AM6234_A53
-	default "AM6234" if SOC_AM6234_M4
-	default "AM6254" if SOC_AM6254_M4 || SOC_AM6254_A53
-	default "AM62L3" if SOC_AM62L3_A53
-	default "AM6442" if SOC_AM6442_M4
-	default "AM6442" if SOC_AM6442_R5F0_0
-	default "AM6442" if SOC_AM6442_R5F0_1
-	default "AM6442" if SOC_AM6442_R5F1_0
-	default "AM6442" if SOC_AM6442_R5F1_1
-	default "J721e" if SOC_J721E_MAIN_R5F0_0
