--- soc/ti/k3/am6x/Kconfig.soc
@@ -1,109 +1,0 @@
-# Copyright (c) 2023 Enphase Energy
-# SPDX-License-Identifier: Apache-2.0
-
-config SOC_SERIES_AM6X
-	bool
-	select SOC_FAMILY_TI_K3
-
-config SOC_SERIES_AM6X_A53
-	bool
-	select SOC_SERIES_AM6X
-	help
-	  Enable support for AM6X A53 Series.
-
-config SOC_SERIES_AM6X_M4
-	bool
-	select SOC_SERIES_AM6X
-	help
-	  Enable support for AM62X M4 Series.
-
-config SOC_SERIES_AM6X_R5
-	bool
-	select SOC_SERIES_AM6X
-	help
-	  Enable support for AM6X R5 Series.
-
-config SOC_AM6234_A53
-	bool
-	select SOC_SERIES_AM6X_A53
-
-config SOC_AM6234_M4
-	bool
-	select SOC_SERIES_AM6X_M4
-
-config SOC_AM6254_A53
-	bool
-	select SOC_SERIES_AM6X_A53
-
-config SOC_AM6254_M4
-	bool
-	select SOC_SERIES_AM6X_M4
-
-config SOC_AM6232_A53
-	bool
-	select SOC_SERIES_AM6X_A53
-
-config SOC_AM6232_M4
-	bool
-	select SOC_SERIES_AM6X_M4
-
-config SOC_AM62L3_A53
-	bool
-	select SOC_SERIES_AM6X_A53
-
-config SOC_AM6442_M4
-	bool
-	select SOC_SERIES_AM6X_M4
-
-config SOC_AM2434_M4
-	bool
-	select SOC_SERIES_AM6X_M4
-
-config SOC_AM2434_R5F0_0
-	bool
-	select SOC_SERIES_AM6X_R5
-
-config SOC_J721E_MAIN_R5F0_0
-	bool
-	select SOC_SERIES_AM6X_R5
-
-config SOC_J722S_MAIN_R5F0_0
-	bool
-	select SOC_SERIES_AM6X_R5
-
-config SOC_J722S_MCU_R5F0_0
-	bool
-	select SOC_SERIES_AM6X_R5
-
-config SOC_AM6442_R5F0_0
-	bool
-	select SOC_SERIES_AM6X_R5
-
-config SOC_AM6442_R5F0_1
-	bool
-	select SOC_SERIES_AM6X_R5
-
-config SOC_AM6442_R5F1_0
-	bool
-	select SOC_SERIES_AM6X_R5
-
-config SOC_AM6442_R5F1_1
-	bool
-	select SOC_SERIES_AM6X_R5
-
-config SOC_SERIES
-	default "am6x" if SOC_SERIES_AM6X
-
-config SOC
-	default "am2434" if SOC_AM2434_M4 || SOC_AM2434_R5F0_0
-	default "am6232" if SOC_AM6232_M4 || SOC_AM6232_A53
-	default "am6234" if SOC_AM6234_M4 || SOC_AM6234_A53
-	default "am6254" if SOC_AM6254_M4 || SOC_AM6254_A53
-	default "am62l3" if SOC_AM62L3_A53
-	default "am6442" if SOC_AM6442_M4
-	default "am6442" if SOC_AM6442_R5F0_0
-	default "am6442" if SOC_AM6442_R5F0_1
-	default "am6442" if SOC_AM6442_R5F1_0
-	default "am6442" if SOC_AM6442_R5F1_1
-	default "j721e" if SOC_J721E_MAIN_R5F0_0
-	default "j722s" if SOC_J722S_MAIN_R5F0_0 || SOC_J722S_MCU_R5F0_0
