+++ soc/ti/k3/common/cortex_r/soc.c
@@ -8,8 +8,8 @@
 #include <zephyr/fatal.h>
 
 #include "soc.h"
+#include "ctrl_partitions.h"
 #include <zephyr/cache.h>
-#include <common/ctrl_partitions.h>
 
 unsigned int z_soc_irq_get_active(void)
 {
