+++ soc/ti/k3/common/cortex_r/soc.h
