+++ soc/ti/k3/common/peripherals/ctrl_partitions.h
