+++ soc/ti/k3/j721e/Kconfig.soc
@@ -1,0 +1,25 @@
+# Copyright (c) 2023 Enphase Energy
+# Copyright (c) 2026 Siemens Mobility GmbH
+# SPDX-License-Identifier: Apache-2.0
+
+config SOC_SERIES_J721E
+	bool
+	select SOC_FAMILY_TI_K3
+
+config SOC_SERIES_J721E_R5
+	bool
+	select SOC_SERIES_J721E
+	help
+	  Enable support for J721E R5 Series.
+
+config SOC_J721E_MAIN_R5F0_0
+	bool
+	select SOC_SERIES_J721E_R5
+
+config SOC_J722S_MAIN_R5F0_0
+	bool
+	select SOC_SERIES_J721E_R5
+
+config SOC_J722S_MCU_R5F0_0
+	bool
+	select SOC_SERIES_J721E_R5
