+++ soc/ti/k3/soc.yml
@@ -1,7 +1,7 @@
 family:
 - name: ti_k3
   series:
-  - name: am6x
+  - name: am62x
     socs:
     - name: am6232
       cpuclusters:
@@ -15,9 +15,13 @@ family:
       cpuclusters:
       - name: m4
       - name: a53
+  - name: am62lx
+    socs:
     - name: am62l3
       cpuclusters:
       - name: a53
+  - name: am64x
+    socs:
     - name: am6442
       cpuclusters:
       - name: m4
@@ -29,9 +33,13 @@ family:
       cpuclusters:
       - name: m4
       - name: r5f0_0
+  - name: j721e
+    socs:
     - name: j721e
       cpuclusters:
       - name: main_r5f0_0
+  - name: j722s
+    socs:
     - name: j722s
       cpuclusters:
       - name: main_r5f0_0
